fix for 64 bit bugs

+++ ViewDeck/IIViewDeckController.m
@@ -1141,7 +1141,7 @@ static NSTimeInterval durationToAnimate(CGFloat pointsToAnimate, CGFloat velocit
         if (_controllers[side] == controller) return side;
     }
     
-    return NSNotFound;
+    return (UInt32)NSNotFound;
 }
